Adds build script and cmake files. Reorganizes files into appropriate folders.
