+++ C/build.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+rm -r bin
+mkdir -p bin
+cd bin
+cmake ..
+make -j$(nproc)
+cd ..
